backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-RESTX==1.3.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2
python-dateutil==2.8.2
pyod==1.1.3
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
openai==1.6.1
Werkzeug==3.0.1
marshmallow==3.20.2
pytest==7.4.3
pytest-flask==1.3.0
python-magic==0.4.27 